refactor: 영역별로 컴포넌트 분리
