+++ src/App.tsx
@@ -1,7 +1,8 @@
+import PostsManagerPage from "@pages/post-manager/ui/PostsManagerPage.tsx"
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header.tsx"
+
 import Footer from "./widgets/ui/Footer.tsx"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import Header from "./widgets/ui/Header.tsx"
 
 const App = () => {
   return (
