+++ tests/basic.test.tsx
@@ -1,12 +1,14 @@
-import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest"
+import "@testing-library/jest-dom"
+
+import PostsManager from "@pages/post-manager/ui/PostsManagerPage"
 import { render, screen, waitFor } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import { http, HttpResponse } from "msw"
 import { setupServer } from "msw/node"
-import { MemoryRouter } from "react-router-dom"
-import PostsManager from "../src/pages/PostsManagerPage"
 import * as React from "react"
-import "@testing-library/jest-dom"
+import { MemoryRouter } from "react-router-dom"
+import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest"
+
 import { TEST_POSTS, TEST_SEARCH_POST, TEST_USERS } from "./mockData"
 
 // MSW 서버 설정
